05-cotacaoDeMoeda/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
Pillow==10.0.1
requests==2.31.0
ttkbootstrap==1.10.1
urllib3==2.0.6
